rename moduleresolver files to tcmoduleresolver for clarity
