+++ lute/luau/CMakeLists.txt
@@ -3,13 +3,13 @@ add_library(Lute.Luau STATIC)
 target_sources(Lute.Luau PRIVATE
     include/lute/configresolver.h
     include/lute/luau.h
-    include/lute/moduleresolver.h
+    include/lute/tcmoduleresolver.h
     include/lute/resolvemodule.h
     include/lute/type.h
 
     src/configresolver.cpp
     src/luau.cpp
-    src/moduleresolver.cpp
+    src/tcmoduleresolver.cpp
     src/resolvemodule.cpp
     src/type.cpp
 )
